Implement automatic OAuth token refresh
The frontend stores the expiration timestamp of the provided access
token and attempts to refresh it if needed by hooking before every API
call. This is transparent for the user and only results in an additional
roundtrip to the API when the access token has expired (or is about to
expire).
Limitations: the token is not refreshed if the API returns a 401, i.e.
if the front does not detect its invalidity based only on its computed
expiration timestamp. This should *not* happen but we should handle this
special case in the future.
Closes #1