Visually indicate offline state
Fade out PlaybackStateView, add an "Offline" tag and rename app title to
"Spotly (offline)" when connection with the API is lost. This can happen
if the API is down or if the user loses Internet connectivity.
The warning is displayed in a non-disturbing way that does not hide
lyrics. The app goes back to normal state once connectivity is restored.
Known limitations: the "Lyrics not found" message gets replaced by the
eventual error.