Add dark mode
Add a dark mode with a discreet toggler in the footer. The toggle
transitions the whole page to dark or light mode. The theme is persisted
across reloads/browser closes (using LocalStorage). The default theme is
light, excepted if the user's browser sets the `prefers-colors-scheme`
value to `dark`. In this case, the theme defaults to dark, but the user
is still able to override it with the toggle button.