cors: cache preflight requests for one hour
Avoid issuing a CORS preflight request (OPTIONS request) before each API
call in the frontend. Browsers cache these requests for 5 seconds but
the frontend calls the API every ~10 seconds. The `maxAge` setting adds
a `Access-Control-Max-Age` header in the CORS response so that browsers
cache the response for the specified time. Settle on 3600s, one hour,
which saves a lot of requests but still allows some flexibility.
Closes #24.